Break out inner class to public class.
