+++ src/Configuration/Configurator.java
@@ -70,46 +70,4 @@ public class Configurator {
         return;
     }
 
-    /**
-     * Static inner class of Configurator. Container for parsed JSON data.
-     */
-    public static class Config {
-
-        /* Instance Variables */
-        private int numFloors;
-        private int numElevators;
-        private long travelTime;
-        private long loadTime;
-
-
-        /* Constructors */
-
-        /**
-         * Default constructor for class Config.
-         */
-        public Config() {
-            this.numFloors = 0;
-            this.numElevators = 0;
-            this.travelTime = 0;
-            this.loadTime = 0;
-        }
-
-        /* Methods */
-        public int getNumFloors() {
-            return numFloors;
-        }
-
-        public int getNumElevators() {
-            return numElevators;
-        }
-
-        public long getTravelTime() {
-            return travelTime;
-        }
-
-        public long getLoadTime() {
-            return loadTime;
-        }
-    }
-
 }
